Refaktorert JS til react
