+++ resourceapp/package.json
@@ -1,16 +1,21 @@
 {
-  "name": "reacttest",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-scripts": "5.0.1",
-    "web-vitals": "^2.1.4"
-  },
+    "name": "reacttest",
+    "version": "0.1.0",
+    "private": true,
+    "options": {
+        "allowedHosts": [ "localhost", ".localhost" ],
+        "proxy": "https://localhost:5000/"
+    },
+    "dependencies": {
+        "@testing-library/jest-dom": "^5.16.5",
+        "@testing-library/react": "^13.4.0",
+        "@testing-library/user-event": "^13.5.0",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-scripts": "5.0.1",
+        "web-vitals": "^2.1.4"
+
+    },
     "scripts": {
         "start": "react-scripts start",
         "build": "react-scripts build",
@@ -22,30 +27,30 @@
         "css:compress": "css-minify -f css/main.css -o css",
         "css:build": "npm-run-all css:compile css:prefix css:compress"
     },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "autoprefixer": "^10.4.13",
-    "concat": "^1.0.3",
-    "css-minify": "^2.0.0",
-    "npm-run-all": "^4.1.5",
-    "postcss-cli": "^10.1.0",
-    "sass": "^1.57.1"
-  }
+    "eslintConfig": {
+        "extends": [
+            "react-app",
+            "react-app/jest"
+        ]
+    },
+    "browserslist": {
+        "production": [
+            ">0.2%",
+            "not dead",
+            "not op_mini all"
+        ],
+        "development": [
+            "last 1 chrome version",
+            "last 1 firefox version",
+            "last 1 safari version"
+        ]
+    },
+    "devDependencies": {
+        "autoprefixer": "^10.4.13",
+        "concat": "^1.0.3",
+        "css-minify": "^2.0.0",
+        "npm-run-all": "^4.1.5",
+        "postcss-cli": "^10.1.0",
+        "sass": "^1.57.1"
+    }
 }
